Use 32bit types

+++ src/event-bridge.cpp
@@ -90,7 +90,7 @@ struct EventBridge::Impl : EventInput::Callback
             input->poll(this);
     }
 
-    bool sendEvent(const EventType etype, const uint8_t index, const int16_t value)
+    bool sendEvent(const EventType etype, const uint8_t index, const int32_t value)
     {
         const uint32_t idx = event_id(etype, index);
 
@@ -106,7 +106,7 @@ struct EventBridge::Impl : EventInput::Callback
 private:
     std::string& last_error;
 
-    void event(const EventType etype, const EventState state, const uint8_t index, const int16_t value) override
+    void event(const EventType etype, const EventState state, const uint8_t index, const int32_t value) override
     {
         if (callback != nullptr)
             callback->eventReceived(etype, state, index, value);
@@ -140,7 +140,7 @@ void EventBridge::poll()
     impl->poll();
 }
 
-bool EventBridge::sendEvent(const EventType etype, const uint8_t index, const int16_t value)
+bool EventBridge::sendEvent(const EventType etype, const uint8_t index, const int32_t value)
 {
     return impl->sendEvent(etype, index, value);
 }
