+++ src/event-bridge.hpp
@@ -22,7 +22,7 @@ struct EventBridge {
         /**
          * Event trigger function, called when an event is received.
          */
-        virtual void eventReceived(EventType etype, EventState state, uint8_t index, int16_t value) = 0;
+        virtual void eventReceived(EventType etype, EventState state, uint8_t index, int32_t value) = 0;
     };
 
    /**
@@ -57,7 +57,7 @@ struct EventBridge {
     /**
      * Event trigger function, to be called for sending events.
      */
-    bool sendEvent(EventType etype, uint8_t index, int16_t value);
+    bool sendEvent(EventType etype, uint8_t index, int32_t value);
 
 private:
     struct Impl;
