+++ src/main.cpp
@@ -149,7 +149,7 @@ struct WebSocketEventBridge : QObject,
     }
 
 private:
-    void eventReceived(EventType etype, EventState state, uint8_t index, int16_t value) override
+    void eventReceived(EventType etype, EventState state, uint8_t index, int32_t value) override
     {
         printf("eventReceived %d:%s, %d:%s, %u, %d\n",
                etype, EventTypeStr(etype), state, EventStateStr(state), index, value);
